refactor: moved dfaCanvas and importing directory into the dfa subdirectory. The exporting isn't dependent on the type of the automaton, but importing is, which is why I moved it into the DFA directory
