+++ finite-automata-designer/public/scripts/dfa/importing/temp.txt
